Adjust test because stat believes permissions
The "-w" test for a file that has been unveiled "r" passes if file
permissions say write should work, which is always the case as root.
I'm unclear why the other permissions don't seem work this way, but not
going to push it.

However, the real test below for opening the file for writing fails
appropritely though.